Merged with andersbll pull request.
In order to support various ground truth geometric models,
the structure of the project had to be slightly changed.
Now the repeatability benchmark constructor takes as first
argument the geometry consistency model object which computes
the ground truth. The API changes slightly as datasets now
return not only a homography but some structure which defines
the geometric model of the scene (differs by model).